Add files via upload
A login page for users who already have an account. The page includes a button to go to the signup.html if they do not have an account. It also contains a button that take the users back to the landing page(index.html) if they do not have an account and if they do not want to create an account